Move cloud.py into __init__.py
